+++ src/command/command.ts
@@ -1,7 +1,7 @@
 import { ExtensionContext, commands } from "vscode";
 
 import { setApiToken, setShowOfflineRobot } from "./globalConfig";
-import { initProject, uploadScript } from "./operation";
+import { initProject, uploadAndRunScript, uploadScript } from "./operation";
 import {
     setProjectName,
     markConfigFile,
@@ -22,7 +22,8 @@ export function registerCommand(context: ExtensionContext): void {
         commands.registerCommand("hamibot-assistant.setExecuteRobot", setExecuteRobot),
 
         // 操作
-        commands.registerCommand("hamibot-assistant.uploadScript", uploadScript),
         commands.registerCommand("hamibot-assistant.initProject", initProject),
+        commands.registerCommand("hamibot-assistant.uploadScript", uploadScript),
+        commands.registerCommand("hamibot-assistant.uploadAndRunScript", uploadAndRunScript),
     );
 }